Adds new search components
* Adds custom element's components fieldset-frame (fieldset for forms),
checkbox and a search components catered to autosuggest kind of text
inputs.
* Changed how the filter prop was being handled in SearchFilter.vue.
There are $emit calls to manipulate the filter prop on the parent
component now.
* Added vue-suggest lib for autosuggests filter, although its UX is
currently not being used (a popup with suggests), its events are.
Also, it's there in case we change our mind ux-wise.
* Added vuebar to option lists to avoid "height overcollapsing"
* Added number-input to control number user input
* Added vue-slider for price component
* Vue's transition groups being used here for a smoother mobile UX
* Common structures got transformed into components themselves:
Icons, SearchFieldset (a fieldset with title, action and slots),
SearchFacetOptions (a list of options), SearchFacetOptionsFilter ( input
 to search a list of options) and SearchFacetPriceFilter ( slider
 to control price value )
* Description of Course is being used as a temporary replacement for
Syllabus