Course page with SSR
Course page is the page where we show and explain more explicitly the
characteristics of a course. It's a glorified offcanvas horizontal
card modal, which means it uses the same Vue components. In fact,
several components were made "shared" because of that.
The Vue app lives on course_page/App.vue and it's comprised of
two sub-components for each Desktop / Mobile version.

Now, since everything is compiled on the client-side, there are two
downsides of building this static-y page using pure Vue: 1) The user
immediately sees a FOUC, especially on slow internet connections,
which is ugly and a terrible UX 2) Google is unable to properly crawl
our website. In order to solve this issue, I decided to use Airbnb's
hypernova project (with its Vue bindings). It's basically a nodejs
server that calls server-side js compilation. The rails side of hypernova
is possible through vueonrails lib. For this server to render what is
originally done on the client side, we have to compile all components
into something nodejs understands. This is why there's a special
webpack.ssr.config. It compiles everything targetting node and generates
a final ssr/ssr.js file, that will be used to run the server (the
configuration of this server lives on initializers/hypernova.rb). The
compilation is done running `yarn ssr` command.

Here's the gist of it all: CoursesController#show renders page using
a special helper `render_vue_component` which calls the nodejs server
running under ssr/ssr.js. This helper also injects props as a script
tag in the html. I also include javascript_pack_tags to allow for Vue
hydration (once the page is rendered on the client-side, Vue picks it
up again and injects itself on the page, reusing the DOM already
generated). The hydration works like this: If ssr was possible, then
there will be a data-hypernova-id div populated with course page content
and then we can use the script tags with props to hydrate. If ssr
wasn't possible, then we simply create a new node on the DOM and inject
Vue ourselves. This makes the app resilient to offline ssr server.

Breakdown of changes:

* To use locales on the client and the server, there's a new i18n.js
file that returns a function creating a VueI18n instance with locale
as param.
* DOMPurify needs special access to js window-based functions so we
have to use a lib to detect whether node or client-side js is running
to decide if we have to mock the window variable.
* Use vue-lazy-hydration so we signal Vue not to touch some DOM nodes
that don't need hydration (nodes that have no interaction)
* SocialSharing vue component is supposed to be used with inline
template syntax. This is why I had to wrap it in a customized shared/
SocialSharing.vue component, with proper template tags.
* Adds conditional loading for base svg sprites. There's no need to
load every sprite on the course page.