Add db structures to generalize signed upload
* Add new "direct" / signed fetch / upload functions
that accept an extra argument for folder. This way,
signed uploads can be scoped on a folder basis
* Direct / signed uploads are always associated with an user
account
* To track different kinds of signed uploads
(for example, provider logos), create a many-to-many relationship
between direct uploads and the targeting table (e.g. app.provider_logos)
* index_provider_logos_on_direct_upload_id: This avoids accidental
insertion of the same direct upload's uuid (might be caused by a bug)
* index_provider_logos_on_provider_id: provider_logo has to be unique
(there's only one logo), although the old logo is kept untouched on the
direct_uploads table
* create a view to manipulate the provider logos: it's a mix between
provider_logos and direct_uploads;
* create a type (app.provider_logo) that is basically the description of
the entry's structure in this view
* this type will be used as a return value for the view_instead
trigger