Style vendors for Classpert and eject webpack
Vendor libraries are now styled and imported from a separate
package called @classpert/vendor-styling. Read the project info
for more information.

For Vue components, the css @import strategy depends on the original
structure of the component. If the component uses scoped CSS, import
the vendor-styling unscoped in a wrapper component that exports the
original component (for example, see NumberInput.vue)

Additional Changes:

Webpack Ejection

Webpacker is no longer used. Refer to the Assets section in the README
file