Add ability to setup course "noindex" strategies
This commit will add the ability to customize "noindex "strategies
per course. A course can be configured to send a special header
`X-Robots-Tag: noindex` to robots, which sends a signal to the robot
to not add this page to its index. This is specially useful for large
web sites (with hundreds of thousands of pages), which have to deal with
limited crawling budgets.
More about the X-Robots-Tag on https://developers.google.com/search/reference/robots_meta_tag