Course Page V2.0 (+ others)
Overhauled course page, this time it's mostly
server side rendered by Rails apart from minor
components that are JS powered like Sharing, Video Preview
and Carrousel that need a bit of JS which means
SSR Vue hydrated in the client being a better option.

Vue components compatible with the newer version of
Elements (9.x family) are versioned with a 2 at the end

To avoid confusion and for remind sake:

CourseSharing is the SSR component which is itself
a wrapper for the CourseSocialSharing component.
The CourseSocialSharing component is a specialization
of either a SocialSharing or SocialSharing2 (GenericSocialSharing
is also a specialization). The SocialSharingX components
are renderless components.

Other changes:

* Added SSR* related components to Carrousel so that
localization only happens during server side rendering
* Updated Swiper version to 5 (which means updated
Awesome Swiper and Vue Awesome Swiper)
* Separated bootstrap from application css so that
there's a sane priority with which css classes are parsed and
applied by the browser
* Fixed slug size problem for Heroku
* Adjusted webpack settings to avoid unnecessary bloat
in production (such as source-map)
* Added constraint to crawler_domains to avoid illegal
transition once the domain is confirmed